Finally done lobby mechanics

+++ Body.py
@@ -38,9 +38,9 @@ start_text = "Игра в 'Перудо'.\n"\
              "  В раунде 'Мапута' единицы становятся ОБЫЧНЫМИ костями и перестают учитыватсья в ходе подсчета кубиков другого номинала, после вскрытия в конце раунда.\n"\
              "  УДАЧИ!"
 
-game_master = ({})
-game_players = ({})
-
+game_master = ({}) # id:name
+game_players = ({}) # -
+id_needed_lobby = []
 game_max_players = 4
 
 
@@ -57,28 +57,40 @@ class Player():
     p_id = 0 #player id
     p_name = "name" #player tg name
     dice_amount = 5
-    dice_value = [""] #contains values of dieces
+    dice_value = [""] #contain diece values
 class Game():
     def __init__(self):
-        self.queue = [] #contains IDs
-        self.queue_names = [] #contains Names
+        self.lobby_id = 1
+        self.queue_id = ({}) # msg_id : lobby_id
+        self.queue_people = {}  #id_player : id_lobby
+        self.queue_people_names = {} # name_player : msg_id
     def add_queue(self, p_id):
         if(p_id in self.queue):
             return False
         else:
             self.queue.append(p_id)
             return True
+    def add_player(self, player_id, needed_lobby):
+        if player_id in self.queue_people:
+            return False
+        else:
+            self.queue_people[f"{player_id}"] = needed_lobby
+            return True
+
+    def create_queue(self, game_master_id, msg_id):
+        if game_master_id not in game_master:
+            self.queue_id[f"{msg_id}"] = self.lobby_id
+            self.lobby_id += 1
 
 g = Game()
 
-###
+###not-for-user-debug
 
 @dp.message(Command("id"))
 async def any_message(message: Message):
     await message.answer(f"{message.from_user.id}")
 
-
-###
+###not-for-user-debug
 
 #KEYBOARD START-----------------------------------------------------------------------------------
 @dp.message(CommandStart())
@@ -100,25 +112,28 @@ async def display_rules(message : Message):
 
 @dp.message(F.text.lower() == "начать")
 async def game_create(message : Message):
-    global game_master
-    if (str(message.from_user.id) not in game_master):
-        game_master[f"{message.from_user.id}"] = message.from_user.first_name
-        i = 4
-        builder = InlineKeyboardBuilder()
-        await message.answer(f"{message.from_user.first_name} - Cоздает лобби игры")
-        for i in range(4,7):
-            builder.add(types.InlineKeyboardButton(
-            text=f"{i}",
-            callback_data=f"p{i}")
-            )
-        builder.row(types.InlineKeyboardButton(text="Продолжить", callback_data="continue"), types.InlineKeyboardButton(text="Удалить лобби", callback_data="distruct")) ###########CALLBACK####################################
-        await message.answer(f"Настройки\nКоличество игроков: {game_max_players}", reply_markup=builder.as_markup())
+    if message.from_user.first_name not in g.queue_people_names:
+        global game_master
+        if (str(message.from_user.id) not in game_master):
+            game_master[f"{message.from_user.id}"] = message.from_user.first_name
+            i = 4
+            builder = InlineKeyboardBuilder()
+            await message.answer(f"{message.from_user.first_name} - Cоздает лобби игры")
+            for i in range(4,7):
+                builder.add(types.InlineKeyboardButton(
+                text=f"{i}",
+                callback_data=f"p{i}")
+                )
+            builder.row(types.InlineKeyboardButton(text="Продолжить", callback_data="continue"), types.InlineKeyboardButton(text="Удалить лобби", callback_data="distruct")) ###########CALLBACK####################################
+            await message.answer(f"Настройки\nКоличество игроков: {game_max_players}", reply_markup=builder.as_markup())
+        else:
+            await message.answer(f"{game_master[f'{message.from_user.id}']}, у вас уже есть активное лобби!")
     else:
-        await message.answer(f"{game_master[f'{message.from_user.id}']}, у вас уже есть активное лобби!")
+        await message.answer(f"{message.from_user.first_name}, к сожалению вы уже присоединились к активному лобби!")
 
 @dp.message(F.text.lower() == "s")
 async def learning(message : Message):
-    #await message.answer(f"{message.chat.id} ТЕКСТ!")
+    #await message.answer(f"{message.from_user.first_name, type(message.from_user.first_name)}")
     await bot.send_message(chat_id='886945416', text="ХУЙ")
 
 
@@ -130,14 +145,32 @@ async def continue_game(callback : types.CallbackQuery):
     builder.row(types.InlineKeyboardButton(text="Открыть Лобби", callback_data="open"), types.InlineKeyboardButton(text="Удалить лобби", callback_data="distruct"))
     await callback.message.answer(f"{game_master[f'{callback.from_user.id}']} - Открывает лобби", reply_markup=builder.as_markup())
 
+
 @dp.callback_query(F.data == "distruct")
 async def delete_lobby(callback : types.CallbackQuery):
     global game_master
-    await callback.message.answer(f"{game_master[f'{callback.from_user.id}']} - Удаляет лобби!")
-    await callback.message.delete()
-    game_master = ({"id": 0, "fst_name": "", "snd_name": ""})
-    game_players = ({"id": 0, "fst_name": "", "snd_name": ""})
-    game_max_players = 4
+    global id_needed_lobby
+    if str(callback.from_user.id) in game_master:
+        await callback.message.answer(f"{game_master[f'{callback.from_user.id}']} - Удаляет лобби!")
+        try:
+            g.queue_id.pop(f"{game_master[f'{callback.from_user.id}']}")
+        except KeyError:
+            pass
+        game_master.pop(f"{callback.from_user.id}")
+        game_players.clear()
+        game_max_players = 4
+        await callback.message.delete()
+    try:
+
+        if str(callback.message.message_id) in g.queue_id:
+            g.queue_id.pop(str(callback.message.message_id))
+        filtered_dict = {}
+        for key, value in g.queue_people_names.items():
+            if str(value) != str(callback.message.message_id):
+                filtered_dict[key] = value
+        g.queue_people_names = filtered_dict
+    except KeyError:
+        pass
 
 @dp.callback_query(F.data == "p4")
 async def p4(callback: types.CallbackQuery):
@@ -180,29 +213,112 @@ async def p6(callback: types.CallbackQuery):
 
 @dp.callback_query(F.data =="open")
 async def display_rules(callback : types.CallbackQuery):
+    try:
+        if callback.from_user.first_name == game_master[f'{callback.from_user.id}']:
+            await callback.message.delete()
+
+            builder = InlineKeyboardBuilder()
+            builder.add(types.InlineKeyboardButton(text="Присоединиться",callback_data='id'))
+            builder.row(types.InlineKeyboardButton(text="Выйти", callback_data='leave'))
+            builder.row(types.InlineKeyboardButton(text="Удалить Лобби", callback_data='distruct'))
+            msg = await callback.message.answer(f"Очередь на игру #{g.lobby_id}", reply_markup=builder.as_markup())
+            g.create_queue(callback.from_user.id, msg.message_id)
+            queue_msg_id = callback.message.message_id
+    except KeyError:
+        pass
+
+@dp.callback_query(F.data == "leave")
+async def leave_lobby(callback : types.CallbackQuery):
+    global game_max_players
     builder = InlineKeyboardBuilder()
     builder.add(types.InlineKeyboardButton(
         text="Присоединиться",
         callback_data='id')
     )
-    await callback.message.answer("Очередь на игру", reply_markup=builder.as_markup())
-    queue_msg_id = callback.message.message_id
+    builder.row(types.InlineKeyboardButton(text="Выйти", callback_data='leave'))
+    builder.row(types.InlineKeyboardButton(text="Удалить Лобби", callback_data='distruct'))
+    value_of_players_in_lobby = -1
+    k = ""
+
+    for i in g.queue_people_names:
+        #await callback.message.answer(f"{g.queue_people_names[i], callback.message.message_id}")
+        if str(g.queue_people_names[i]) == str(callback.message.message_id):
+            value_of_players_in_lobby += 1
+    key = callback.from_user.first_name
+    value = g.queue_people_names.pop(str(callback.from_user.first_name))
+    for key, value in g.queue_people_names.items():
+        if str(value) == str(callback.message.message_id):
+            k += key + ', '
+
+    await callback.answer(text=f"{callback.from_user.first_name} - выходит из лобби!", cache_time=5, show_alert=True)
+    await bot.edit_message_text(chat_id=callback.message.chat.id, message_id=value,text=f"Очередь на игру({value_of_players_in_lobby}/{game_max_players}):\n{k}",reply_markup=builder.as_markup())
+    value_of_players_in_lobby = 0
+    k = ""
+    if value_of_players_in_lobby == 0:
+        if str(callback.from_user.id) in game_master:
+            await callback.message.answer(f"{game_master[f'{callback.from_user.id}']} - Удаляет лобби!")
+            try:
+                g.queue_id.pop(f"{game_master[f'{callback.from_user.id}']}")
+            except KeyError:
+                pass
+            game_master.pop(f"{callback.from_user.id}")
+            game_players.clear()
+            game_max_players = 4
+            await callback.message.delete()
+        try:
+
+            if str(callback.message.message_id) in g.queue_id:
+                g.queue_id.pop(str(callback.message.message_id))
+            filtered_dict = {}
+            for key, value in g.queue_people_names.items():
+                if str(value) != str(callback.message.message_id):
+                    filtered_dict[key] = value
+            g.queue_people_names = filtered_dict
+        except KeyError:
+            pass
+
 @dp.callback_query(F.data == "id")
 async def send_message(callback: types.CallbackQuery):
-    if len(g.queue_names) <= game_max_players:
-        if(g.add_queue(callback.from_user.id) == True):
-            g.queue_names.append(callback.from_user.first_name)
+    global id_needed_lobby
+    builder = InlineKeyboardBuilder()
+    builder.add(types.InlineKeyboardButton(
+        text="Присоединиться",
+        callback_data='id')
+    )
+    builder.row(types.InlineKeyboardButton(text="Выйти", callback_data='leave'))
+    builder.row(types.InlineKeyboardButton(text="Удалить Лобби", callback_data='distruct'))
+    id_needed_lobby.clear()
+    for i in g.queue_id:
+        if i in g.queue_id:
+            id_needed_lobby.append(g.queue_id[i])
+            id_needed_lobby.append(i)
+
+    g.queue_people[f'{callback.from_user.id}'] = id_needed_lobby[0]
+
+    value_of_players_in_lobby = 0
+
+    if value_of_players_in_lobby <= game_max_players:
+        if (g.add_player(callback.from_user.id, id_needed_lobby[len(id_needed_lobby)-1]) == True):
+            #g.queue_people_names[callback.from_user.first_name] = str(id_needed_lobby[id_needed_lobby[len(id_needed_lobby)-1]])
+            g.queue_people_names[f"{callback.from_user.first_name}"] = id_needed_lobby[len(id_needed_lobby) - 1]
             await callback.message.answer(f"{callback.from_user.first_name}, вы встали в очередь на игру")
         else:
             await callback.message.answer(f"{callback.from_user.first_name}, Вы уже стоите в очереди на игру")
-        builder = InlineKeyboardBuilder()
-        builder.add(types.InlineKeyboardButton(
-            text="Присоединиться",
-            callback_data='id')
-        )
-        await bot.edit_message_text(chat_id=callback.message.chat.id, message_id=callback.message.message_id, text=f"Очередь на игру({len(g.queue_names)}/{game_max_players}):\n{g.queue_names}",reply_markup=builder.as_markup())
     else:
         await callback.message.answer('Очередь заполнена')
+    k = ''
+    #for i in list(g.queue_people_names.keys()):
+    #    k += i + ", "
+    for key, value in g.queue_people_names.items():
+        if value == id_needed_lobby[len(id_needed_lobby)-1]:
+            k += key + ', '
+    #await callback.message.answer(f'{id_needed_lobby[len(id_needed_lobby)-1]}')
+    for i in g.queue_people_names:
+        if g.queue_people_names[i] == id_needed_lobby[len(id_needed_lobby)-1]:
+            value_of_players_in_lobby += 1
+    #await bot.send_message(text=f"{g.queue_people_names} INL", chat_id=callback.message.chat.id)
+    await bot.edit_message_text(chat_id=callback.message.chat.id, message_id=id_needed_lobby[len(id_needed_lobby)-1],text=f"Очередь на игру({value_of_players_in_lobby}/{game_max_players}):\n{k}",reply_markup=builder.as_markup())
+
 #KEYBOARD START-----------------------------------------------------------------------------------
 
 if __name__ == "__main__":
